requirements.txt
mistralai
python-dotenv
rich
